+++ src/components/DexCoverage.module.css
@@ -1,0 +1,4 @@
+.DexCoverage_root {
+  font-variant-numeric: tabular-nums;
+  max-width: 16rem;
+}
